refactor: `products` API module을 `features/product`로 분리
